Add settings file to enable compatibility with VertexHeightOblateAdvanced.
see settings.cfg for more information.

The default setting is Classic, which will basically be a slightly prettier version of what we're all used to. ClassicTexture is the original method using a heightmap. PointEquipotentialGrav is the only setting which is currently is compatible with gameplay, but is not thoroughly tested.

VertexHeightOblateAdvanced can be found here: https://github.com/jamespglaze/VertexHeightOblateAdvanced/releases