asteroid improvements
fix asteroid generation in a few obscure cases.
Diddmun should now generate, but the current problem is that it will generate endlessly. This may not be fixable.

NB: this patch contains spoilers, do please do not inspect too closely if you don't want something really cool to be potentially spoiled.